Elimina el archivo de preguntas generadas y el script de extracción de PDF. Crea un nuevo script de extracción de PDF que procesa archivos y guarda resultados en JSON. Agrega imágenes para las materias y actualiza el dashboard para mostrar materias registradas con un nuevo diseño y funcionalidad de botón.
