--- Datos/preguntas_generadas.json
@@ -1,132 +1,0 @@
-[
-  {
-    "pregunta": "¿Qué tipo de datos se pueden considerar como no estructurados?",
-    "opciones": [
-      "A) Texto natural",
-      "B) Imágenes",
-      "C) Audio",
-      "D) Todos los anteriores"
-    ],
-    "respuesta_correcta": "D) Todos los anteriores",
-    "explicacion": "Los datos no estructurados son aquellos que no tienen una forma organizada o predefinida, como texto natural, imágenes, audio, video, entre otros.",
-    "dificultad": "baja",
-    "tema": "Tipos de datos"
-  },
-  {
-    "pregunta": "¿Cuál es el objetivo principal de la minería de datos?",
-    "opciones": [
-      "A) Extraer conocimiento",
-      "B) Predecir patrones",
-      "C) Entrenar modelos",
-      "D) Analizar tendencias"
-    ],
-    "respuesta_correcta": "A) Extraer conocimiento",
-    "explicacion": "La minería de datos se enfoca en extraer conocimiento valioso y relevante de grandes conjuntos de datos, lo que puede ser utilizado para tomar decisiones informadas.",
-    "dificultad": "media",
-    "tema": "Minería de datos"
-  },
-  {
-    "pregunta": "¿Qué tipo de algoritmo se utiliza comúnmente en la segmentación de clientes?",
-    "opciones": [
-      "A) Clustering",
-      "B) Regresión logística",
-      "C) K-NN",
-      "D) Random Forest"
-    ],
-    "respuesta_correcta": "A) Clustering",
-    "explicacion": "La segmentación de clientes es un proceso que utiliza algoritmos de aprendizaje automático, como el clustering, para agrupar a los clientes en categorías basadas en sus características y comportamientos.",
-    "dificultad": "media",
-    "tema": "Aprendizaje automático"
-  },
-  {
-    "pregunta": "¿Qué es un modelo predictivo?",
-    "opciones": [
-      "A) Un modelo que describe la relación entre variables",
-      "B) Un modelo que predice el futuro con base en patrones pasados",
-      "C) Un modelo que identifica outliers en los datos",
-      "D) Un modelo que clasifica objetos"
-    ],
-    "respuesta_correcta": "B) Un modelo que predice el futuro con base en patrones pasados",
-    "explicacion": "Un modelo predictivo se utiliza para predecir futuras observaciones o resultados con base en patrones y tendencias identificadas en los datos.",
-    "dificultad": "media",
-    "tema": "Modelos predictivos"
-  },
-  {
-    "pregunta": "¿Qué tipo de análisis se enfoca en la relación entre variables?",
-    "opciones": [
-      "A) Análisis descriptivo",
-      "B) Análisis predicativo",
-      "C) Análisis correlacional",
-      "D) Análisis causativo"
-    ],
-    "respuesta_correcta": "C) Análisis correlacional",
-    "explicacion": "El análisis correlacional se enfoca en la relación entre variables, identificando patrones y tendencias que permiten comprender cómo las variables están relacionadas.",
-    "dificultad": "baja",
-    "tema": "Análisis estadístico"
-  },
-  {
-    "pregunta": "¿Cuál es el objetivo principal de la ciencia de datos?",
-    "opciones": [
-      "A) Recopilar grandes cantidades de datos",
-      "B) Analizar y visualizar datos para obtener insights",
-      "C) Aplicar técnicas estadísticas para predecir resultados",
-      "D) Crear modelos predictivos para tomar decisiones"
-    ],
-    "respuesta_correcta": "B) Analizar y visualizar datos para obtener insights",
-    "explicacion": "La ciencia de datos busca extraer conocimiento e insights de los datos para informar decisiones y mejorar procesos.",
-    "dificultad": "baja",
-    "tema": "Introducción a la ciencia de datos"
-  },
-  {
-    "pregunta": "¿Qué tipo de datos se consideran no estructurados?",
-    "opciones": [
-      "A) Datos numéricos y categóricos",
-      "B) Texto, imágenes y audio",
-      "C) Datos geoespaciales",
-      "D) Datos temporales"
-    ],
-    "respuesta_correcta": "B) Texto, imágenes y audio",
-    "explicacion": "Los datos no estructurados son aquellos que no tienen una forma predefinida o ordenada, como texto, imágenes, audio y video.",
-    "dificultad": "media",
-    "tema": "Tipos de datos"
-  },
-  {
-    "pregunta": "¿Qué algoritmo se utiliza comúnmente para clasificar objetos en categorías?",
-    "opciones": [
-      "A) Algoritmo de k-means",
-      "B) Algoritmo de árbol de decisión",
-      "C) Algoritmo de redes neuronales",
-      "D) Algoritmo de cluster"
-    ],
-    "respuesta_correcta": "B) Algoritmo de árbol de decisión",
-    "explicacion": "El algoritmo de árbol de decisión es un tipo de aprendizaje automático que se utiliza comúnmente para clasificar objetos en categorías.",
-    "dificultad": "media",
-    "tema": "Aprendizaje automático"
-  },
-  {
-    "pregunta": "¿Qué tipo de análisis estadístico se utiliza para identificar patrones y tendencias en los datos?",
-    "opciones": [
-      "A) Análisis de series temporales",
-      "B) Análisis de regresión lineal",
-      "C) Análisis de varianza",
-      "D) Análisis de correlación"
-    ],
-    "respuesta_correcta": "A) Análisis de series temporales",
-    "explicacion": "El análisis de series temporales se utiliza para identificar patrones y tendencias en los datos sobre el tiempo.",
-    "dificultad": "media",
-    "tema": "Análisis estadístico"
-  },
-  {
-    "pregunta": "¿Qué herramienta de visualización de datos es popularmente utilizada?",
-    "opciones": [
-      "A) Tableau",
-      "B) Power BI",
-      "C) Excel",
-      "D) Google Data Studio"
-    ],
-    "respuesta_correcta": "A) Tableau",
-    "explicacion": "Tableau es una herramienta de visualización de datos que se utiliza comúnmente para crear informes y dashboards.",
-    "dificultad": "baja",
-    "tema": "Herramientas de visualización"
-  }
-]
